Corrección de errores del repositorio. Además, se eliminaron de la clase Movie de dominio las represetaciones propias de la capa de datos (remoteMoviem ,...).
