+++ composeApp/src/desktopMain/kotlin/edu/dyds/movies/data/local/DataEntities.kt
@@ -1,8 +1,7 @@
-package edu.dyds.movies.data
+package edu.dyds.movies.data.local
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import edu.dyds.movies.domain.Movie
 
 @Serializable
 data class RemoteMovie(
@@ -16,20 +15,7 @@ data class RemoteMovie(
     @SerialName("original_language") val originalLanguage: String,
     val popularity: Double,
     @SerialName("vote_average") val voteAverage: Double
-) {
-    fun toDomain(): Movie = Movie(
-        id = id,
-        title = title,
-        overview = overview,
-        releaseDate = releaseDate,
-        poster = "https://image.tmdb.org/t/p/w185$posterPath",
-        backdrop = backdropPath?.let { "https://image.tmdb.org/t/p/w780$it" },
-        originalTitle = originalTitle,
-        originalLanguage = originalLanguage,
-        popularity = popularity,
-        voteAverage = voteAverage
-    )
-}
+)
 
 @Serializable
 data class RemoteResult(
