+++ composeApp/src/desktopMain/kotlin/edu/dyds/movies/data/local/DataMapper.kt
@@ -1,0 +1,16 @@
+package edu.dyds.movies.data.local // o data.mapper si preferís
+
+import edu.dyds.movies.domain.entity.Movie
+
+fun RemoteMovie.toDomain(): Movie = Movie(
+    id = id,
+    title = title,
+    overview = overview,
+    releaseDate = releaseDate,
+    poster = "https://image.tmdb.org/t/p/w185$posterPath",
+    backdrop = backdropPath?.let { "https://image.tmdb.org/t/p/w780$it" },
+    originalTitle = originalTitle,
+    originalLanguage = originalLanguage,
+    popularity = popularity,
+    voteAverage = voteAverage
+)
