+++ composeApp/src/desktopMain/kotlin/edu/dyds/movies/domain/repository/MoviesRepository.kt
@@ -1,9 +1,9 @@
 package edu.dyds.movies.domain.repository
 
-import edu.dyds.movies.domain.entity.RemoteMovie
+import edu.dyds.movies.data.local.RemoteMovie
 
 interface MoviesRepository {
     //quedarian mejor los nombres sin TMDB
     suspend fun getPopularMovies(): List<RemoteMovie>
-    suspend fun getMovieDetails(id: Int): RemoteMovie
+    suspend fun getMovieDetails(id: Int): RemoteMovie?
 }
